LayoutTests/imported/w3c/web-platform-tests/custom-elements/registries/CustomElementRegistry-upgrade-expected.txt


PASS upgrade is a function on both global and scoped CustomElementRegistry
PASS upgrade is a no-op when called on a shadow root with no association
PASS upgrade is a no-op when called on an element associated with a different registry
PASS upgrade should upgrade a candidate element when called on a shadow root with an association
PASS upgrade should not upgrade a candidate element not associated with a registry
